added product edit page

+++ src/Components/Pages/My Toys/MyToys.jsx
@@ -1,23 +1,57 @@
-import React, { useState } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
 import { AiFillDelete } from 'react-icons/ai';
 import { BiEdit } from 'react-icons/bi';
-import { useLoaderData } from 'react-router-dom';
+import { useLoaderData, useNavigate } from 'react-router-dom';
+import Swal from 'sweetalert2';
+import { AuthDetials } from '../../Providers/AuthProviders';
 
 const MyToys = () => {
-    const loadedData = useLoaderData();
-    const [products, setProducts] = useState(loadedData);
+    // const loadedData = useLoaderData();
+    const {user} = useContext(AuthDetials);
+    const email = user.email;
+    const [products, setProducts] = useState([]);
     const [slice, setSlice] = useState(8);
+    const navigate = useNavigate();
+
+    useEffect(() => {
+        const loadingData = async () => {
+            const responce = await fetch(`http://localhost:3000/my-toys?email=${email}`)
+            const data = await responce.json();
+            setProducts(data)
+        }
+        loadingData();
+    })
     const handleSlice = () => {
         setSlice(slice + 5)
     }
+    const handleEdit = (id) => {
+        navigate(`/edit/${id}`)
+    }
     const handleDelete = (id) => {
-        fetch(`http://localhost:3000/product/${id}`, {
-            method: "DELETE",
-        }).then((res) => res.json())
-        .then((data) => {
-            const remaining = products.filter(product => product._id !== id);
-            setProducts(remaining);
-        })
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:3000/product/${id}`, {
+                    method: "DELETE",
+                }).then((res) => res.json())
+                .then((data) => {
+                    const remaining = products.filter(product => product._id !== id);
+                    setProducts(remaining);
+                })
+              Swal.fire(
+                'Deleted!',
+                'Your file has been deleted.',
+                'success'
+              )
+            }
+          })
     }
     return (
         <div className="overflow-x-auto">
@@ -47,7 +81,8 @@ const MyToys = () => {
                                     <td>
                                         <div className='flex text-4xl gap-4'>
                                             <BiEdit
-                                                className='text-green-600 cursor-pointer' />
+                                            onClick={() => handleEdit(product._id)}
+                                            className='text-green-600 cursor-pointer' />
                                             <AiFillDelete 
                                             onClick={() => handleDelete(product._id)}
                                             className='text-red-600 cursor-pointer' />
